requirements.txt
cassandra_driver
Flask
pandas
scikit_learn
Unidecode
xgboost
